Prevents `GeventWebSocketClient.send` from blocking on full queues.
If one wishes to have a process which stores references to a number
of connections, and then sends out messages to them based on events,
then the maximum queue length is problematic, since it means that
if any messages have not been sent for a particular connection,
then any future send calls to that connection will block until
the first message has been sent. Given that in the case of an empty
queue `send` does not provide any guarantee of how quickly the
message will be sent, it doesn't provide any benefit for subsequent
calls to be blocked based on prior unsent messages.