exhaust receive buffer if available, upstream #48
Equivalent to upstream pull request #48 in asyncio.
Once websocket_recv_ub is called reader or recv_event would be
cancelled and only the first message would be returned. The other
messages would be left in buffer till new coming packet triggers reader
or recv_event.
This could hardly happen if rate is low. But if user is flooding, this
happens. Once flood stopped, the last few messages would probably be
left.
Call websocket_recv_ub continuously till an empty message received
would be a solution but with flaw, because exhausted buffer is not the
only reason of receiving an empty message. Since uwsgi non-blocking
websocket api do NOT distinguish a really empty message and blocking
occasion, please notice that if user is flooding messages with empty
message mixed, this would still happen.